separate tests for the record API in their own file
move tests from `record.rs`, which was getting bulky, into a separate
file
