+++ rust/template/differential_datalog/record.rs
@@ -5,23 +5,18 @@ use std::vec;
 use std::collections::{btree_map, BTreeMap, BTreeSet};
 use std::iter::FromIterator;
 use std::borrow::Cow;
-#[cfg(test)]
-use std::borrow;
 use std::ptr::{null_mut, null};
 use std::ffi::CStr;
 use std::os::raw::c_char;
 use libc::size_t;
 use std::fmt;
 
-#[cfg(test)]
-use num::bigint::{ToBigInt, ToBigUint};
-
 pub type Name = Cow<'static, str>;
 
 /// `enum Record` represents an arbitrary DDlog value.
 ///
-/// It relies on string to store constructor and field names.  When manufacturing an instance of
-/// `Record` from a typed DDlog, strings can be cheap `&'static str`'s.  When manufacturing an
+/// It relies on strings to store constructor and field names.  When manufacturing an instance of
+/// `Record` from a typed DDlog value, strings can be cheap `&'static str`'s.  When manufacturing an
 /// instance from some external representation, e.g., JSON, one needs to use `String`'s instead.  To
 /// accommodate both options, `Record` uses `Cow` to store names.
 ///
@@ -44,18 +39,15 @@ pub enum CollectionKind {
     Map
 }
 
-
+/// Relation can be identified by name (e.g., when parsing JSON or text)
+/// or ID, which is more efficient if the caller bothered to convert
+/// relation name to ID.
 #[derive(Debug,PartialEq,Eq,Clone)]
 pub enum RelIdentifier {
     RelName(Name),
     RelId(usize)
 }
 
-/*
- * Relation can be identified by name (e.g., when parsing JSON or text)
- * or ID, which is more efficient if the caller bothered to convert
- * relation name to ID.
- */
 impl fmt::Display for RelIdentifier {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
@@ -65,6 +57,7 @@ impl fmt::Display for RelIdentifier {
     }
 }
 
+/// Four types of DDlog relation update commands that match the `Update` enum in `program.rs`
 #[derive(Debug,PartialEq,Eq,Clone)]
 pub enum UpdCmd {
     Insert (RelIdentifier, Record),
@@ -76,7 +69,6 @@ pub enum UpdCmd {
 /*
  * C API to Record and UpdCmd
  */
-
 
 #[no_mangle]
 pub unsafe extern "C" fn ddlog_free(rec: *mut Record)
@@ -193,7 +185,6 @@ pub unsafe extern "C" fn ddlog_get_tuple_size(rec: *const Record) -> size_t {
     }
 }
 
-
 #[no_mangle]
 pub unsafe extern "C" fn ddlog_get_tuple_field(rec: *const Record, idx: size_t) -> *const Record {
     match rec.as_ref() {
@@ -201,7 +192,6 @@ pub unsafe extern "C" fn ddlog_get_tuple_field(rec: *const Record, idx: size_t) 
         _ => null_mut()
     }
 }
-
 
 /// Convenience method to create a 2-tuple.
 #[no_mangle]
@@ -243,7 +233,6 @@ pub unsafe extern "C" fn ddlog_get_vector_size(rec: *const Record) -> size_t {
         _ => 0
     }
 }
-
 
 #[no_mangle]
 pub unsafe extern "C" fn ddlog_get_vector_elem(rec: *const Record, idx: size_t) -> *const Record {
@@ -318,7 +307,6 @@ pub unsafe extern "C" fn ddlog_is_map(rec: *const Record) -> bool {
         _ => false
     }
 }
-
 
 #[no_mangle]
 pub unsafe extern "C" fn ddlog_get_map_size(rec: *const Record) -> size_t {
@@ -453,6 +441,10 @@ pub unsafe extern "C" fn ddlog_delete_key_cmd(table: libc::size_t, rec: *mut Rec
     Box::into_raw(Box::new(UpdCmd::DeleteKey(RelIdentifier::RelId(table), *rec)))
 }
 
+/*
+ * Traits for converting ddlog `Values` to/from Record's.
+ */
+
 /// `Mutator` trait represents an object that can be used to mutate a value (e.g., change some of
 /// its fields).
 pub trait Mutator<V> {
@@ -468,6 +460,10 @@ pub trait FromRecord: Sized {
 pub trait IntoRecord {
     fn into_record(self) -> Record;
 }
+
+/*
+ * Trait implementations for built-in types.
+ */
 
 impl FromRecord for u8 {
     fn from_record(val: &Record) -> Result<Self, String> {
@@ -498,15 +494,6 @@ impl IntoRecord for u8 {
     }
 }
 
-#[test]
-fn test_u8() {
-    assert_eq!(u8::from_record(&Record::Int(25_u8.to_bigint().unwrap())), Ok(25));
-    assert_eq!(u8::from_record(&Record::Int(0xab.to_bigint().unwrap())), Ok(0xab));
-    assert_eq!(u8::from_record(&Record::Int(0xabcd.to_bigint().unwrap())), Err("cannot convert 43981 to u8".to_string()));
-    assert_eq!(u8::into_record(0x25), Record::Int(BigInt::from(0x25)));
-}
-
-
 impl FromRecord for u16 {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -535,15 +522,6 @@ impl Mutator<u16> for Record {
         Ok(())
     }
 }
-
-#[test]
-fn test_u16() {
-    assert_eq!(u16::from_record(&Record::Int(25_u16.to_bigint().unwrap())), Ok(25));
-    assert_eq!(u16::from_record(&Record::Int(0xab.to_bigint().unwrap())), Ok(0xab));
-    assert_eq!(u16::from_record(&Record::Int(0xabcdef.to_bigint().unwrap())), Err("cannot convert 11259375 to u16".to_string()));
-    assert_eq!(u16::into_record(32000), Record::Int(BigInt::from(32000)));
-}
-
 
 impl FromRecord for u32 {
     fn from_record(val: &Record) -> Result<Self, String> {
@@ -574,15 +552,6 @@ impl Mutator<u32> for Record {
     }
 }
 
-#[test]
-fn test_u32() {
-    assert_eq!(u32::from_record(&Record::Int(25_u32.to_bigint().unwrap())), Ok(25));
-    assert_eq!(u32::from_record(&Record::Int(0xab.to_bigint().unwrap())), Ok(0xab));
-    assert_eq!(u32::from_record(&Record::Int(0xabcdef.to_bigint().unwrap())), Ok(0xabcdef));
-    assert_eq!(u32::from_record(&Record::Int((-0xabcdef).to_bigint().unwrap())), Err("cannot convert -11259375 to u32".to_string()));
-}
-
-
 impl FromRecord for u64 {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -610,14 +579,6 @@ impl Mutator<u64> for Record {
         *v = u64::from_record(self)?;
         Ok(())
     }
-}
-
-#[test]
-fn test_u64() {
-    assert_eq!(u64::from_record(&Record::Int(25_u64.to_bigint().unwrap())), Ok(25));
-    assert_eq!(u64::from_record(&Record::Int(0xab.to_bigint().unwrap())), Ok(0xab));
-    assert_eq!(u64::from_record(&Record::Int(0xabcdef.to_bigint().unwrap())), Ok(0xabcdef));
-    assert_eq!(u64::from_record(&Record::Int((-0xabcdef).to_bigint().unwrap())), Err("cannot convert -11259375 to u64".to_string()));
 }
 
 impl FromRecord for u128 {
@@ -649,16 +610,6 @@ impl Mutator<u128> for Record {
     }
 }
 
-#[test]
-fn test_u128() {
-    assert_eq!(u128::from_record(&Record::Int(25_u128.to_bigint().unwrap())), Ok(25));
-    assert_eq!(u128::from_record(&Record::Int(100000000000000000000000000000000000000_u128.to_bigint().unwrap())), Ok(100000000000000000000000000000000000000));
-    assert_eq!(u128::from_record(&Record::Int(0xab.to_bigint().unwrap())), Ok(0xab));
-    assert_eq!(u128::from_record(&Record::Int(0xabcdef.to_bigint().unwrap())), Ok(0xabcdef));
-    assert_eq!(u128::from_record(&Record::Int((-0xabcdef).to_bigint().unwrap())), Err("cannot convert -11259375 to u128".to_string()));
-}
-
-
 impl FromRecord for BigInt {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -681,12 +632,6 @@ impl Mutator<BigInt> for Record {
         *v = BigInt::from_record(self)?;
         Ok(())
     }
-}
-
-#[test]
-fn test_bigint() {
-    let v = (-25_i64).to_bigint().unwrap();
-    assert_eq!(BigInt::from_record(&Record::Int(v.clone())), Ok(v));
 }
 
 impl FromRecord for BigUint {
@@ -718,17 +663,6 @@ impl Mutator<BigUint> for Record {
     }
 }
 
-#[test]
-fn test_biguint() {
-    let vi = (25_i64).to_bigint().unwrap();
-    let vu = (25_i64).to_biguint().unwrap();
-    assert_eq!(BigUint::from_record(&Record::Int(vi)), Ok(vu));
-
-    let vi = (-25_i64).to_bigint().unwrap();
-    assert_eq!(BigUint::from_record(&Record::Int(vi)), Err("cannot convert -25 to BigUint".to_string()));
-}
-
-
 impl FromRecord for bool {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -753,12 +687,6 @@ impl Mutator<bool> for Record {
     }
 }
 
-#[test]
-fn test_bool() {
-    assert_eq!(bool::from_record(&Record::Bool(true)), Ok(true));
-}
-
-
 impl FromRecord for String {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -781,12 +709,6 @@ impl Mutator<String> for Record {
         *v = String::from_record(self)?;
         Ok(())
     }
-}
-
-#[test]
-fn test_string() {
-    assert_eq!(String::from_record(&Record::String("foo".to_string())), Ok("foo".to_string()));
-    assert_eq!(String::from_record(&Record::Bool(true)), Err("not a string Bool(true)".to_string()));
 }
 
 impl FromRecord for () {
@@ -860,33 +782,6 @@ decl_tuple_from_record!(18, T0, 0, T1, 1, T2, 2, T3, 3, T4, 4, T5, 5, T6, 6, T7,
 decl_tuple_from_record!(19, T0, 0, T1, 1, T2, 2, T3, 3, T4, 4, T5, 5, T6, 6, T7, 7, T8, 8, T9, 9, T10, 10, T11, 11, T12, 12, T13, 13, T14, 14, T15, 15, T16, 16, T17, 17, T18, 18);
 decl_tuple_from_record!(20, T0, 0, T1, 1, T2, 2, T3, 3, T4, 4, T5, 5, T6, 6, T7, 7, T8, 8, T9, 9, T10, 10, T11, 11, T12, 12, T13, 13, T14, 14, T15, 15, T16, 16, T17, 17, T18, 18, T19, 19);
 
-#[test]
-fn test_tuple() {
-    assert_eq!(<(bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false)])),
-               Ok((true, false)));
-    assert_eq!(<(bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false)])),
-               Ok((true, false, false)));
-    assert_eq!(<(bool, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true)])),
-               Ok((true, false, false, true)));
-    assert_eq!(<(bool, bool, bool, bool, String)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string())])),
-               Ok((true, false, false, true, "foo".to_string())));
-    assert_eq!(<(bool, bool, bool, bool, String, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false)])),
-               Ok((true, false, false, true, "foo".to_string(), false)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false), Record::Bool(false)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false, false)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false), Record::Bool(false), Record::Bool(true)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false, false, true)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::Bool(true)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false, false, true, true)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool, bool, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::Bool(true), Record::Bool(true)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false, false, true, true, true)));
-    assert_eq!(<(bool, bool, bool, bool, String, bool, bool, bool, bool, bool, bool, bool)>::from_record(&Record::Tuple(vec![Record::Bool(true), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::String("foo".to_string()), Record::Bool(false), Record::Bool(false), Record::Bool(false), Record::Bool(true), Record::Bool(true), Record::Bool(true), Record::Bool(false)])),
-               Ok((true, false, false, true, "foo".to_string(), false, false, false, true, true, true, false)));
-}
-
-
 impl<T: FromRecord> FromRecord for vec::Vec<T> {
     fn from_record(val: &Record) -> Result<Self, String> {
         match val {
@@ -912,12 +807,6 @@ impl<T: FromRecord> Mutator<vec::Vec<T>> for Record {
         *v = <vec::Vec<T>>::from_record(self)?;
         Ok(())
     }
-}
-
-#[test]
-fn test_vec() {
-    assert_eq!(<vec::Vec<bool>>::from_record(&Record::Array(CollectionKind::Unknown, vec![Record::Bool(true), Record::Bool(false)])),
-               Ok(vec![true, false]));
 }
 
 macro_rules! decl_arr_from_record {
@@ -986,15 +875,6 @@ decl_arr_from_record!(30);
 decl_arr_from_record!(31);
 decl_arr_from_record!(32);
 
-#[test]
-fn test_array() {
-    assert_eq!(<[bool;2]>::from_record(&Record::Array(CollectionKind::Vector, vec![Record::Bool(true), Record::Bool(false)])),
-               Ok([true, false]));
-    assert_eq!(<[bool;2]>::from_record(&Record::Array(CollectionKind::Vector, vec![Record::Bool(true)])),
-               Err("cannot convert Array(Vector, [Bool(true)]) to array of length 2".to_owned()));
-}
-
-
 impl<K: FromRecord+Ord, V: FromRecord> FromRecord for BTreeMap<K,V> {
     fn from_record(val: &Record) -> Result<Self, String> {
         vec::Vec::from_record(val).map(|v|BTreeMap::from_iter(v))
@@ -1033,41 +913,6 @@ impl<K: FromRecord+Ord, V: FromRecord + PartialEq> Mutator<BTreeMap<K,V>> for Re
     }
 }
 
-#[test]
-fn test_map() {
-    assert_eq!(<BTreeMap<u32, u32>>::from_record(&Record::Array(CollectionKind::Unknown,
-                                                                vec![Record::Tuple(vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(10))]),
-                                                                     Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ])),
-               Ok(BTreeMap::from_iter(vec![(0,10), (1,10)])));
-
-    let mut v = <BTreeMap<u32, u32>>::from_record(&Record::Array(CollectionKind::Unknown,
-                                                  vec![Record::Tuple(vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(10))]),
-                                                       Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ])).unwrap();
-    Record::Array(CollectionKind::Unknown,
-                  vec![Record::Tuple(vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(10))]),
-                       Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ]).
-               mutate(&mut v).unwrap();
-    assert_eq!(v, BTreeMap::from_iter(vec![]));
-
-    v = <BTreeMap<u32, u32>>::from_record(&Record::Array(CollectionKind::Unknown,
-                                          vec![Record::Tuple(vec![Record::Int(BigInt::from(2)), Record::Int(BigInt::from(20))]),
-                                               Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(20))] ) ])).unwrap();
-    Record::Array(CollectionKind::Unknown,
-                             vec![Record::Tuple(vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(10))]),
-                                  Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ]).
-               mutate(&mut v).unwrap();
-    assert_eq!(v, BTreeMap::from_iter(vec![(0,10), (1,10), (2,20)]));
-
-    v =  <BTreeMap<u32, u32>>::from_record(&Record::Array(CollectionKind::Unknown,
-                                            vec![Record::Tuple(vec![Record::Int(BigInt::from(2)), Record::Int(BigInt::from(20))]),
-                                                 Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ])).unwrap();
-    Record::Array(CollectionKind::Unknown,
-                             vec![Record::Tuple(vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(10))]),
-                                  Record::Tuple(vec![Record::Int(BigInt::from(1)), Record::Int(BigInt::from(10))] ) ]).
-               mutate(&mut v).unwrap();
-    assert_eq!(v, BTreeMap::from_iter(vec![(0,10), (2,20)]));
-}
-
 impl<T: FromRecord+Ord> FromRecord for BTreeSet<T> {
     fn from_record(val: &Record) -> Result<Self, String> {
         vec::Vec::from_record(val).map(|v|BTreeSet::from_iter(v))
@@ -1094,25 +939,12 @@ impl<T: FromRecord+Ord> Mutator<BTreeSet<T>> for Record {
     }
 }
 
-#[test]
-fn test_set() {
-    assert_eq!(<BTreeSet<u32>>::from_record(&Record::Array(CollectionKind::Unknown, vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(1))])),
-               Ok(BTreeSet::from_iter(vec![0, 1])));
-
-    let mut v = <BTreeSet<u32>>::from_record(&Record::Array(CollectionKind::Unknown,
-                                             vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(2))])).unwrap();
-    Record::Array(CollectionKind::Unknown, vec![Record::Int(BigInt::from(0)), Record::Int(BigInt::from(1))]).
-               mutate(&mut v).unwrap();
-    assert_eq!(v, BTreeSet::from_iter(vec![1, 2]));
-}
-
-
-
-#[cfg(test)]
-#[derive(Eq, PartialEq, Debug, Clone, Default)]
-struct Foo<T> {
-    f1: T
-}
+/*
+ * Use the following macros to generate `IntoRecord` and `Mutator` trait implementations for
+ * user-defined structs and enums.
+ *
+ * TODO: implement similar macros for `FromRecord`.
+ */
 
 pub fn arg_extract<T:FromRecord+Default>(args: &Vec<(Name, Record)>, argname: &str) -> Result<T, String> {
     args.iter().find(|(n,_)|*n==argname).map_or_else(||Ok(Default::default()), |(_,v)| T::from_record(v))
@@ -1122,33 +954,6 @@ pub fn arg_find<'a>(args: &'a Vec<(Name, Record)>, argname: &str) -> Option<&'a 
     args.iter().find(|(n,_)|*n==argname).map(|(_,v)|v)
 }
 
-
-#[cfg(test)]
-impl <T: FromRecord+Default> FromRecord for Foo<T> {
-    fn from_record(val: &Record) -> Result<Self, String> {
-        match val {
-            Record::PosStruct(constr, args) => {
-                match constr.as_ref() {
-                    "Foo" if args.len() == 1 => {
-                        Ok(Foo{f1: T::from_record(&args[0])?})
-                    },
-                    c => Result::Err(format!("unknown constructor {} of type Foo in {:?}", c, *val))
-                }
-            },
-            Record::NamedStruct(constr, args) => {
-                match constr.as_ref() {
-                    "Foo" => {
-                        Ok(Foo{f1: arg_extract(args, "f1")?})
-                    },
-                    c => Result::Err(format!("unknown constructor {} of type Foo in {:?}", c, *val))
-                }
-            },
-            v => {
-                Result::Err(format!("not a struct {:?}", *v))
-            }
-        }
-    }
-}
 
 #[macro_export]
 macro_rules! decl_struct_into_record {
@@ -1185,104 +990,6 @@ macro_rules! decl_record_mutator_struct {
             }
         }
     };
-}
-
-#[cfg(test)]
-pub struct NestedStruct<T>{x:bool, y: Foo<T>}
-
-#[cfg(test)]
-pub struct StructWithNoFields;
-
-#[cfg(test)]
-decl_struct_into_record!(Foo, <T>, f1);
-
-#[cfg(test)]
-decl_record_mutator_struct!(Foo, <T>, f1: T);
-
-#[cfg(test)]
-decl_struct_into_record!(NestedStruct, <T>, x,y);
-
-#[cfg(test)]
-decl_record_mutator_struct!(NestedStruct, <T>, x: bool, y: Foo<T>);
-
-#[cfg(test)]
-decl_struct_into_record!(StructWithNoFields, <>,);
-
-#[cfg(test)]
-decl_record_mutator_struct!(StructWithNoFields, <>, );
-
-#[test]
-fn test_struct()
-{
-    let foo1: Foo<BTreeMap<u32, String>> = Foo{f1: BTreeMap::from_iter(vec![(5,"five".to_owned()), (6,"six".to_owned())])};
-    let foo2 = <Foo<BTreeMap<u32, String>>>::from_record(&foo1.clone().into_record()).unwrap();
-
-    assert_eq!(foo1, foo2);
-
-    let upd = Record::NamedStruct(Cow::from("Foo"),
-                                  vec![(Cow::from("f1"), Record::Array(CollectionKind::Unknown,
-                                                                       vec![Record::Tuple(vec![Record::Int(BigInt::from(5)), Record::String("5".to_owned())])]))]);
-    let mut foo_mod = foo1.clone();
-    upd.mutate(&mut foo_mod).unwrap();
-    let foo_expected = Foo{f1: BTreeMap::from_iter(vec![(5,"5".to_owned()), (6,"six".to_owned())])};
-    assert_eq!(foo_mod, foo_expected);
-}
-
-
-#[cfg(test)]
-type Bbool = bool;
-
-#[cfg(test)]
-#[derive(Eq, PartialEq, Debug, Clone)]
-enum DummyEnum<T> {
-    Constr1{f1: Bbool, f2: String},
-    Constr2{f1: T, f2: BigInt, f3: Foo<T>},
-    Constr3{f1: (bool, bool)},
-}
-
-#[cfg(test)]
-impl <T: FromRecord+Default> FromRecord for DummyEnum<T> {
-    fn from_record(val: &Record) -> Result<Self, String> {
-        match val {
-            Record::PosStruct(constr, args) => {
-                match constr.as_ref() {
-                    "Constr1" if args.len() == 2 => {
-                        Ok(DummyEnum::Constr1{f1: <Bbool>::from_record(&args[0])?,
-                                              f2: String::from_record(&args[1])?})
-                    },
-                    "Constr2" if args.len() == 3 => {
-                        Ok(DummyEnum::Constr2{f1: <T>::from_record(&args[0])?,
-                                              f2: <BigInt>::from_record(&args[1])?,
-                                              f3: <Foo<T>>::from_record(&args[2])?})
-                    },
-                    "Constr3" if args.len() == 1 => {
-                        Ok(DummyEnum::Constr3{f1: <(bool,bool)>::from_record(&args[0])?})
-                    },
-                    c => Result::Err(format!("unknown constructor {} of type DummyEnum in {:?}", c, *val))
-                }
-            },
-            Record::NamedStruct(constr, args) => {
-                match constr.as_ref() {
-                    "Constr1" if args.len() == 2 => {
-                        Ok(DummyEnum::Constr1{f1: arg_extract::<Bbool>(args, "f1")?,
-                                              f2: arg_extract::<String>(args, "f2")?})
-                    },
-                    "Constr2" if args.len() == 3 => {
-                        Ok(DummyEnum::Constr2{f1: arg_extract::<T>(args, "f1")?,
-                                              f2: arg_extract::<BigInt>(args, "f2")?,
-                                              f3: arg_extract::<Foo<T>>(args, "f3")?})
-                    },
-                    "Constr3" if args.len() == 1 => {
-                        Ok(DummyEnum::Constr3{f1: arg_extract::<(bool,bool)>(args, "f1")?})
-                    },
-                    c => Result::Err(format!("unknown constructor {} of type DummyEnum in {:?}", c, *val))
-                }
-            },
-            v => {
-                Result::Err(format!("not a struct {:?}", *v))
-            }
-        }
-    }
 }
 
 #[macro_export]
@@ -1344,44 +1051,6 @@ macro_rules! decl_enum_into_record {
         }
     };
 }
-
-#[cfg(test)]
-decl_enum_into_record!(DummyEnum,<T>,Constr1{f1,f2},Constr2{f1,f2,f3},Constr3{f1});
-
-#[cfg(test)]
-decl_record_mutator_enum!(DummyEnum,<T>,Constr1{f1:Bbool ,f2: String},Constr2{f1: T, f2: BigInt, f3: Foo<T>},Constr3{f1: (bool, bool)});
-
-#[test]
-fn test_enum()
-{
-    assert_eq!(DummyEnum::from_record(&Record::PosStruct(Cow::from("Constr1"),
-                                                    vec![Record::Bool(true), Record::String("foo".to_string())])),
-               Ok(DummyEnum::Constr1::<bool>{f1: true, f2: "foo".to_string()}));
-    assert_eq!(DummyEnum::from_record(&Record::NamedStruct(Cow::from("Constr1"),
-                                                    vec![(Cow::from("f1"), Record::Bool(true)), (Cow::from("f2"), Record::String("foo".to_string()))])),
-               Ok(DummyEnum::Constr1::<bool>{f1: true, f2: "foo".to_string()}));
-    assert_eq!(DummyEnum::from_record(&Record::PosStruct(Cow::from("Constr2"),
-                                                    vec![Record::Int((5_i64).to_bigint().unwrap()),
-                                                         Record::Int((25_i64).to_bigint().unwrap()),
-                                                         Record::PosStruct(Cow::from("Foo"), vec![Record::Int((0_i64).to_bigint().unwrap())])])),
-               Ok(DummyEnum::Constr2::<u16>{f1: 5,
-                                            f2: (25_i64).to_bigint().unwrap(),
-                                            f3: Foo{f1: 0}}));
-    assert_eq!(DummyEnum::from_record(&Record::NamedStruct(Cow::from("Constr2"),
-                                                    vec![(Cow::from("f1"), Record::Int((5_i64).to_bigint().unwrap())),
-                                                         (Cow::from("f2"), Record::Int((25_i64).to_bigint().unwrap())),
-                                                         (Cow::from("f3"), Record::NamedStruct(Cow::from("Foo"),
-                                                                                                vec![(Cow::from("f1"), Record::Int((0_i64).to_bigint().unwrap()))]))])),
-               Ok(DummyEnum::Constr2::<u16>{f1: 5,
-                                            f2: (25_i64).to_bigint().unwrap(),
-                                            f3: Foo{f1: 0}}));
-
-    let enm = DummyEnum::Constr2::<u16>{f1: 5,
-                                        f2: (25_i64).to_bigint().unwrap(),
-                                        f3: Foo{f1: 0}};
-    assert_eq!(DummyEnum::from_record(&enm.clone().into_record()), Ok(enm));
-}
-
 
 #[macro_export]
 macro_rules! decl_val_enum_into_record {
